feat: make activity indicator reusable
